Refactor try_run
